ci: enhance Docker image build and push workflow
Refactor the GitHub Actions workflow for building and pushing the Docker image.
This update introduces:
- Use of `docker/setup-buildx-action` for improved build capabilities.
- `docker/metadata-action` to automatically generate comprehensive image tags and labels based on branch, commit SHA, date, and semantic versioning.
- `docker/build-push-action` for streamlined build and push operations, including caching.

This significantly improves the automation and metadata associated with the Docker image builds.

build: update Dockerfile for Symfony and security checker

- Change Symfony CLI installer download method from `wget` to `curl`.
- Add a step to download and create a Symfony project using `curl`.
- Update `local-php-security-checker` download to use `curl`.

docs: update README with new tools

Add `ffmpeg` to the list of tools in the README.